感度登録完了時にアラートが出るように修正
